+++ ruitk/themes/fallen/staging.js
@@ -4,7 +4,7 @@ import { Convert } from "../../support/convert.js";
 import { Tester } from "../../support/tester.js";
 
 let elements = [
-    {
+    { // search
         name: "search",
         function: (info, element) => {
             info = Merge.dicts({
@@ -37,8 +37,8 @@ let elements = [
             }
             e.addEventListener("keydown", (event) => handleKeyDown(event, info.callback));
             
-            let icon = element.makeElements(`<icon>{name : "editFind", color : "var(--accent3)", c : 2, r : 2}`);
-
+            let icon = element.makeElements(`<icon>{name : "editFind", color : "var(--accent1)",
+            hover_color : "var(--accent2)", size : "var(--fontSizeH2)", c : 2, r : 2}`);
             function handleClick(callback) {
                 const input = document.getElementById(element.parse(info.id));
                 const lastValue = input.dataset.lastValue || "";
@@ -68,6 +68,10 @@ let elements = [
                 size: "var(--fontSizeP1)",  // Default size
                 color: "var(--standout4)", // Default color
             }, info);
+
+            info = Merge.dicts({
+                hoverColor : info.color,
+            }, info);
             
             Style.fontFace({
                 fontFamily : "icons",
@@ -81,7 +85,7 @@ let elements = [
             let e = element.generate(info, element);
             e.innerHTML = Convert.convert(info.name, "dashedCase")
             e.style.fontSize = info.size;
-            e.style.color = info.color;
+            e = Style.style(e, [{color : info.color, hover_color : info.hoverColor}, element.style]);
     
             return e;
         },
@@ -97,9 +101,10 @@ let elements = [
             lineHeight: "1",
         },
         style_standard : "<base>",
-        element: "i"
+        element: "i",
+        handleStyle : true,
     },
-    {
+    { // img
         name: "img",
         function: (info, element) => {
             info = Merge.dicts({
@@ -115,12 +120,49 @@ let elements = [
         },
         style_standard : "<base>",
     },
-    {
+    { // markdown
         name: "markdown",
         function: (info, element) => {
             info = Merge.dicts({
                 str : "",
             }, info);
+
+            let elements = [];
+
+            function findElement(str, pattern, type) {
+                let matches = [];
+                let match;
+                
+                while ((match = pattern.exec(str)) !== null) {
+                    matches.push({ str : match[0], content : match[1], index : match.index, type : type});
+                }
+                
+                return matches;
+            }
+
+            /**
+             * please note: This is intend to work the same as obsidain during edge cases
+             * with the following exemptions
+             *  - no incomplete statements Eg *italic
+             */
+            
+
+            // headings
+            elements.concat(findMatches(info.str, /^#\s+(.+)$/gm, "h1"));
+            elements.concat(findMatches(info.str, /^##\s+(.+)$/gm, "h2"));
+            elements.concat(findMatches(info.str, /^###\s+(.+)$/gm, "h3"));
+            elements.concat(findMatches(info.str, /^####\s+(.+)$/gm, "h4"));
+            elements.concat(findMatches(info.str, /^#####\s+(.+)$/gm, "h5"));
+            elements.concat(findMatches(info.str, /^######\s+(.+)$/gm, "h6"));
+
+            elements.concat(findMatches(info.str, /^(.+)\n=+$/gm, "h1"));
+            elements.concat(findMatches(info.str, /^(.+)\n--+$/gm, "h2"));
+
+            // bold
+            elements.concat(findMatches(info.str, /\*\*([^\*]+)\*\*/gm, "b"));
+            elements.concat(findMatches(info.str, /\s__(.+?)__\s/gm, "b"));
+
+            // italic
 
             
         },